DS: Cleanup API to dispose inlay components.
``createComponent``'s "disposable" argument is useless because it points to the inlay which can be reused (and not disposed).

Current approach is to make component ``Disposable``, but regular swing ``JComponent`` can't be make ``Disposable`` in JVM without of redundant inheritance, and can't be delegated because it is not an interface.

This fix provides API to return separate ``JComponent`` and ``Disposable`` and also removes useless "disposable" argument

GitOrigin-RevId: 127fc31e20aa73920a5740a0785213d307762d49